Refactor jbh_ringtone plugin to use JbhRingtoneModel for ringtone data. Added RingtoneListPage for displaying ringtones with filtering options. Updated method signatures for compatibility with new model structure.
